order can be viewed in detail and status can be updated

+++ SDPCore/Models/DataModel/OrderLine.cs
@@ -13,14 +13,14 @@ namespace Microsoft.SDP.SDPCore.Models
         public Guid lineId { get; set; }
         public Order order { get; set; }
         public Product product { get; set; }
-        public int quantity { get; set; }
+        public decimal quantity { get; set; }
         public decimal discount { get; set; }
 
         public OrderLine()
         {
         }
 
-        public OrderLine( Order order, Product product, int quantity, decimal discount)
+        public OrderLine( Order order, Product product, decimal quantity, decimal discount)
         {
             this.lineId = Guid.NewGuid();
             this.order = order;
